Разбил utils на несколько файлов
