+++ src/view/site-menu.js
@@ -1,5 +1,5 @@
 import AbstractView from "./abstract.js";
-import {upperFirst} from "../utils.js";
+import {upperFirst} from "../utils/common.js";
 
 const createFilterItemTemplate = ({name, count}, isActive) => {
